Add worker pool to crackZipPass
This commit adds a basic worker pool to the crackZipPass tool.

I've attempted to keep the code as much as possible as is, although there's a bit of a flaw with the way passwords are checked. It's possible for a ZIP archive to have different
passwords for each file, in the area that the password is being checked we do indeed check over all files for the given password, however we exit once the first password is found.

I may come back and visit this for you later but it will require substantial changes from the original and will make it quite a bit more complicated.

Before adding a worker pool and moving the password to the end of the file:
[+] Found password
[+] Password = 111111

real 0m38.083s
user 0m38.324s
sys 0m0.223s

And after:
[+] Found password
[+] Password = 111111

real 0m0.408s
user 0m0.663s
sys 0m0.363s